+++ src/integTest/groovy/com/monits/gradle/sca/FindbugsIntegTest.groovy
@@ -143,6 +143,7 @@ class FindbugsIntegTest extends AbstractPerSourceSetPluginIntegTestFixture {
         given:
         writeBuildFile()
         useEmptySuppressionFilter()
+        //noinspection LongLine
         file('src/main/java/com/monits/ClassA.java') << '''
             package com.monits;
 
@@ -243,14 +244,15 @@ class FindbugsIntegTest extends AbstractPerSourceSetPluginIntegTestFixture {
         gradleVersion = gradleVersionForAndroid(androidVersion)
     }
 
-    @SuppressWarnings('MethodName')
-    void 'multimodule android project has all classes'() {
+    @SuppressWarnings(['MethodName', 'LineLength', 'LongLine'])
+    @Unroll('multimodule classes are available when using android gradle plugin #androidVersion and gradle #gradleVersion')
+    void 'multimodule classes are available'() {
         given:
-        setupMultimoduleAndroidProject()
+        setupMultimoduleAndroidProject(androidVersion)
 
         when:
         BuildResult result = gradleRunner()
-                .withGradleVersion(gradleVersionForAndroid(DEFAULT_ANDROID_VERSION))
+                .withGradleVersion(gradleVersion)
                 .build()
 
         then:
@@ -263,6 +265,58 @@ class FindbugsIntegTest extends AbstractPerSourceSetPluginIntegTestFixture {
 
         // make sure nothing is reported
         finbugsReport.assertContents(containsString('<Errors errors="0" missingClasses="0">'))
+
+        where:
+        androidVersion << AndroidLintIntegTest.ANDROID_PLUGIN_VERSIONS
+        gradleVersion = gradleVersionForAndroid(androidVersion)
+    }
+
+    @Unroll('AAR classes are available when using android gradle plugin #androidVersion and gradle #gradleVersion')
+    @SuppressWarnings('MethodName')
+    void 'Dependency AAR classes are available'() {
+        given:
+        writeAndroidBuildFile(androidVersion) << '''
+            dependencies {
+                compile 'io.card:android-sdk:5.5.1'
+            }
+
+            android {
+                defaultConfig {
+                    minSdkVersion 16
+                }
+            }
+        '''
+        writeAndroidManifest()
+        useEmptySuppressionFilter()
+        file('src/main/java/com/monits/ClassA.java') << '''
+            package com.monits;
+
+            import io.card.payment.CardIOActivity;
+
+            public class ClassA {
+                public boolean isFoo() {
+                    return new CardIOActivity().isFinishing();
+                }
+            }
+        '''
+
+        when:
+        BuildResult result = gradleRunner()
+            .withGradleVersion(gradleVersion)
+            .build()
+
+        then:
+        if (GradleVersion.version(gradleVersion) >= GradleVersion.version('2.5')) {
+            result.task(taskName()).outcome == SUCCESS
+        }
+
+        // The report must exist, and not complain on missing classes from liba
+        reportFile().exists()
+        reportFile().assertContents(containsString('<Errors errors="0" missingClasses="0">'))
+
+        where:
+        androidVersion << AndroidLintIntegTest.ANDROID_PLUGIN_VERSIONS
+        gradleVersion = gradleVersionForAndroid(androidVersion)
     }
 
     @SuppressWarnings('MethodName')